test for scrut token

+++ epab/cmd/_pytest.py
@@ -86,7 +86,9 @@ class _CoverageConfigFile:
         Returns:
 
         """
-        if os.getenv('SCRUT_TOK', False):
+        test = os.environ
+        test2 = os.getenv('SCRUT_TOK')
+        if os.getenv('SCRUT_TOK', False) and Path('.coverage.xml').exists():
             epab.utils.AV.info('uploading coverage to Scrutinizer')
             epab.utils.run('pip install scrutinizer-ocular')
             token = os.getenv('SCRUT_TOK')
